refactor(codegen)!: parse CLI with clap (#297)
Replace the handwritten antlr4-rust-gen parser, help, version, and value-enum handling with clap derives now that codegen is isolated in its own crate. Enable clap's derive, env, and cargo features while keeping clap out of the runtime dependency graph.

Drop the ANTLR-style single-dash listener and visitor aliases in favor of standard double-dash options. Move CompilerConfig out of the CLI module so Builder and generation internals remain independent of command-line parsing.

Declare single-consumer third-party dependencies in their owning crate manifests, while retaining shared and synchronized workspace-package dependencies at the workspace root. Update CLI snapshots, rejection coverage, documentation, and the changelog.

+++ crates/antlr-rust-codegen/src/parser/decision.rs
@@ -13,7 +13,7 @@ struct DecisionAltLook {
 /// Upper bound accepted by `--fixed-lookahead`. Dispatch-table size and
 /// analysis cost grow with depth; beyond a few tokens the tier stops paying
 /// for itself, so cap the flag rather than let a typo explode generation.
-pub(crate) const MAX_FIXED_LOOKAHEAD_FLAG: usize = 8;
+pub(crate) const MAX_FIXED_LOOKAHEAD_FLAG: u8 = 8;
 /// Per-alternative cap on lookahead rectangles gathered by the fixed-LL(k)
 /// walk. This is an *analysis* budget: generous enough that ordinary
 /// decisions get an honest disjoint / not-disjoint verdict (fan-outy
